chore(protocol): Refactor Block Info Txs
